+++ src/coral_fans/CoralFans.cpp
@@ -11,6 +11,7 @@
 #include "coral_fans/functions/prof/Prof.h"
 #include "coral_fans/functions/shortcuts/Shortcuts.h"
 #include "coral_fans/functions/slime/Slime.h"
+#include "coral_fans/functions/tick/TickCommandManager.h"
 #include "coral_fans/functions/village/Village.h"
 #include "ll/api/Config.h"
 #include "ll/api/event/EventBus.h"
@@ -99,6 +100,8 @@ void CoralFans::unhook() {
     functions::NoclipManager::getInstance().hook(false);
     functions::hookTick(false, true);
     functions::CFVillageManager::hookVillage(false);
+    functions::miningHook(false);
+    functions::TickCommandManager::getInstance().hook(false);
 
     // for (auto& eventListener : getEventListeners()) ll::event::EventBus::getInstance().removeListener(eventListener);
     getEventListeners().clear();
@@ -106,6 +109,7 @@ void CoralFans::unhook() {
 
 void CoralFans::removeRuntimeData() {
     functions::FreeCameraManager::getInstance().FreeCamList.clear();
+    functions::TickCommandManager::getInstance().clear();
     functions::HopperCounterManager::getInstance().clearAllData();
     functions::HsaManager::getInstance().setHsaShow(false);
     functions::HsaManager::getInstance().setStructureShow(false);