+++ src/coral_fans/commands/Tick.cpp
@@ -1,7 +1,9 @@
 #include "Commands.h"
 #include "coral_fans/CoralFans.h"
+#include "coral_fans/base/Macros.h"
 #include "coral_fans/base/MySchedule.h"
 #include "coral_fans/base/Utils.h"
+#include "coral_fans/functions/tick/TickCommandManager.h"
 
 
 #include "ll/api/command/CommandHandle.h"
@@ -17,6 +19,7 @@
 #include "mc/server/commands/CommandRegistry.h"
 #include "mc/util/Timer.h"
 #include "mc/world/Minecraft.h"
+#include "mc/world/actor/player/Player.h"
 #include "mc/world/level/Level.h"
 
 
@@ -44,16 +47,18 @@ void registerTickCommand(config::CommandConfigStruct& config) {
         .execute([&](CommandOrigin const&, CommandOutput& output, ll::command::RuntimeCommand const& self) {
             using ll::i18n_literals::operator""_tr;
             const auto val = self["tickFreezeType"].get<ll::command::ParamKind::Enum>();
-            // LevelEventPacket{LevelEvent::SimTimeStep, origin.getWorldPosition(), pause}.sendToClients();
-            auto mc = ll::service::getMinecraft();
-            if (mc.has_value()) {
-                if (val.index) mc->setSimTimePause(true);
-                else {
+            auto       mc  = ll::service::getMinecraft();
+            if (mc.has_value()) [[likely]] {
+                if (val.index) {
+                    mc->setSimTimePause(true);
+                    output.success("command.tick.set.freeze"_tr(val.name));
+                } else {
                     mc->setSimTimePause(false);
                     mc->setSimTimeScale(1.0f);
+                    functions::TickCommandManager::getInstance().applyReset();
+                    output.success("command.tick.set.reset"_tr(val.name));
                 }
-            }
-            output.success("command.tick.set.output"_tr(val.name));
+            } else return output.error("command.tick.rate.error.generic"_tr());
         });
 
     // tick rate <float>
@@ -63,13 +68,33 @@ void registerTickCommand(config::CommandConfigStruct& config) {
         .execute([](CommandOrigin const&, CommandOutput& output, ll::command::RuntimeCommand const& self) {
             using ll::i18n_literals::operator""_tr;
             float rate = self["rate"].get<ll::command::ParamKind::Float>();
-            if (rate < 0) output.error("command.tick.rate.error"_tr());
-            // LevelEventPacket{LevelEvent::SimTimeScale, {rate / 20}, rate > 0}.sendToClients();
+            if (rate < 0) {
+                output.error("command.tick.rate.error.outofrange"_tr(0.0f, rate));
+                return;
+            }
             auto mc = ll::service::getMinecraft();
+            if (!mc.has_value()) [[unlikely]]
+                return output.error("command.tick.rate.error.generic"_tr());
 
             mc->setSimTimePause(false);
-            if (mc.has_value()) mc->setSimTimeScale(rate / 20.0f);
+            mc->setSimTimeScale(rate / 20.0f);
+            functions::TickCommandManager::getInstance().applyRateChange(rate / 20.0f);
             output.success("command.tick.rate.success"_tr(rate));
+        });
+
+    // tick sync [bool]
+    tickCommand.runtimeOverload()
+        .text("sync")
+        .optional("enable", ll::command::ParamKind::Bool)
+        .execute([](CommandOrigin const& origin, CommandOutput& output, ll::command::RuntimeCommand const& self) {
+            using ll::i18n_literals::operator""_tr;
+            COMMAND_CHECK_PLAYER
+            auto& manager = functions::TickCommandManager::getInstance();
+            bool  enable  = self["enable"].has_value() ? self["enable"].get<ll::command::ParamKind::Bool>()
+                                                       : !manager.isSynced(player->getRealName());
+            manager.setSync(*player, enable);
+            output.success("command.tick.sync.output"_tr(enable ? "true" : "false"));
+            if (enable) utils::segmentAndSendToPlayer("§e" + "command.tick.sync.warning"_tr(), player);
         });
 
     // tick query [int]
@@ -89,20 +114,17 @@ void registerTickCommand(config::CommandConfigStruct& config) {
             }
             auto uuid = player.value() ? player.value()->getUuid().asString() : "";
             my_schedule::MySchedule::getSchedule().add([uuid, tick](int&, int& count) {
+                auto message = "command.tick.query.mspt"_tr(
+                    ProfilerLite::gProfilerLiteInstance().mDebugServerTickTime->count() / 1000000.0
+                );
+                if (auto mc = ll::service::getMinecraft(); mc && mc->mSimTimer.mSteppingTick > 1e-4) {
+                    message += "command.tick.query.targetmspt"_tr(1000.0f / mc->mSimTimer.mTimeScale * 20);
+                }
                 if (uuid != "") {
                     auto player = ll::service::getLevel()->getPlayer(mce::UUID(uuid));
-                    if (player)
-                        utils::segmentAndSendToPlayer(
-                            "command.tick.query.output"_tr(
-                                ProfilerLite::gProfilerLiteInstance().mDebugServerTickTime->count() / 1000000.0
-                            ),
-                            player
-                        );
+                    if (player) utils::segmentAndSendToPlayer(message, player);
                     else return false; // 玩家不在线了，停止任务
-                } else
-                    CoralFans::getInstance().getSelf().getLogger().info("command.tick.query.output"_tr(
-                        ProfilerLite::gProfilerLiteInstance().mDebugServerTickTime->count() / 1000000.0
-                    ));
+                } else CoralFans::getInstance().getSelf().getLogger().info(message);
 
                 count++;
                 return tick > count;
@@ -120,8 +142,13 @@ void registerTickCommand(config::CommandConfigStruct& config) {
                 return;
             }
             auto mc = ll::service::getMinecraft();
-            if (mc.has_value()) mc->mSimTimer.mSteppingTick = (float)tick;
-            output.success("command.tick.step.output"_tr(tick));
+            if (mc.has_value()) [[likely]] {
+                mc->mSimTimer.mSteppingTick = (float)tick;
+                functions::TickCommandManager::getInstance().applyReset();
+                output.success("command.tick.step.output"_tr(tick));
+            } else return output.error("command.tick.rate.error.generic"_tr());
         });
+
+    functions::TickCommandManager::getInstance().hook(true);
 }
 } // namespace coral_fans::commands