+++ src/coral_fans/functions/freeCamera/FreeCamera.cpp
@@ -6,6 +6,7 @@
 #include "mc/legacy/ActorUniqueID.h"
 #include "mc/network/ServerNetworkHandler.h"
 #include "mc/network/packet/AddPlayerPacket.h"
+#include "mc/network/packet/PlayerAuthInputPacket.h"
 #include "mc/network/packet/PlayerSkinPacket.h"
 #include "mc/network/packet/RemoveActorPacket.h"
 #include "mc/network/packet/UpdateAbilitiesPacket.h"
@@ -91,7 +92,7 @@ void SendActorLinkPacket(Player* pl) {
 */
 
 void FreeCameraManager::EnableFreeCamera(Player* pl) {
-    FreeCameraManager::getInstance().FreeCamList.insert(pl->getNetworkIdentifier().mGuid.g);
+    FreeCameraManager::getInstance().FreeCamList.insert(pl->getUuid().asString());
     EnableFreeCameraPacket(pl);
     SendFakePlayerPacket(pl);
     // SendActorLinkPacket(pl);
@@ -101,7 +102,7 @@ void FreeCameraManager::DisableFreeCamera(Player* pl) {
     auto pos   = pl->getFeetPos();
     auto dimid = pl->getDimensionId();
     // auto links = pl->getLinks();
-    FreeCameraManager::getInstance().FreeCamList.erase(pl->getNetworkIdentifier().mGuid.g);
+    FreeCameraManager::getInstance().FreeCamList.erase(pl->getUuid().asString());
     DisableFreeCameraPacket(pl);
     pl->teleport(pos, dimid);
     // for (auto& link : links) {
@@ -119,12 +120,13 @@ LL_TYPE_INSTANCE_HOOK(
     ServerNetworkHandler,
     &ServerNetworkHandler::$handle,
     void,
-    NetworkIdentifier const&     id,
-    PlayerAuthInputPacket const& pkt
+    NetworkIdentifier const&     source,
+    PlayerAuthInputPacket const& packet
 ) {
-    RETURN_IF_NOT_MAIN_THREAD(return origin(id, pkt));
-    if (!FreeCameraManager::getInstance().FreeCamList.contains(id.mGuid.g)) [[likely]] {
-        origin(id, pkt);
+    RETURN_IF_NOT_MAIN_THREAD(return origin(source, packet));
+    auto player = thisFor<NetEventCallback>()->_getServerPlayer(source, packet.mSenderSubId);
+    if (!FreeCameraManager::getInstance().FreeCamList.contains(player->getUuid().asString())) [[likely]] {
+        origin(source, packet);
     }
 }
 
@@ -138,7 +140,7 @@ LL_TYPE_INSTANCE_HOOK(
 ) {
     RETURN_IF_NOT_MAIN_THREAD(return origin(gamemode));
     origin(gamemode);
-    if (FreeCameraManager::getInstance().FreeCamList.contains(getNetworkIdentifier().mGuid.g)) {
+    if (FreeCameraManager::getInstance().FreeCamList.contains(getUuid().asString())) {
         FreeCameraManager::DisableFreeCamera(this);
     }
 }
@@ -156,7 +158,7 @@ LL_TYPE_INSTANCE_HOOK(
     RETURN_IF_NOT_MAIN_THREAD(return origin(source, damage, hurtParameters));
     auto res = origin(source, damage, hurtParameters);
     if ((this->isSurvival() || this->isAdventure())
-        && FreeCameraManager::getInstance().FreeCamList.contains(this->getNetworkIdentifier().mGuid.g)) {
+        && FreeCameraManager::getInstance().FreeCamList.contains(getUuid().asString())) {
         FreeCameraManager::DisableFreeCamera(this);
     }
     return res;
@@ -171,7 +173,7 @@ LL_TYPE_INSTANCE_HOOK(
     class ActorDamageSource const& a1
 ) {
     RETURN_IF_NOT_MAIN_THREAD(return origin(a1));
-    if (FreeCameraManager::getInstance().FreeCamList.contains(getNetworkIdentifier().mGuid.g)) {
+    if (FreeCameraManager::getInstance().FreeCamList.contains(getUuid().asString())) {
         FreeCameraManager::DisableFreeCamera(this);
     }
     return origin(a1);
@@ -185,7 +187,7 @@ LL_TYPE_INSTANCE_HOOK(
     void
 ) {
     RETURN_IF_NOT_MAIN_THREAD(return origin());
-    FreeCameraManager::getInstance().FreeCamList.erase(getNetworkIdentifier().mGuid.g);
+    FreeCameraManager::getInstance().FreeCamList.erase(getUuid().asString());
     return origin();
 }