+++ src/coral_fans/functions/func/ContainerOpen.cpp
@@ -50,17 +50,18 @@ LL_TYPE_INSTANCE_HOOK(
     GameMode,
     &GameMode::$useItemOn,
     InteractionResult,
-    ItemStack&      item,
-    BlockPos const& blockPos,
-    uchar           face,
-    Vec3 const&     clickPos,
-    Block const*    block,
-    bool            isFirstEvent
+    ::ItemStack&      item,
+    ::BlockPos const& at,
+    uchar             face,
+    ::Vec3 const&     hit,
+    ::HandSlot        handSlot,
+    ::Block const*    targetBlock,
+    bool              isFirstEvent
 ) {
-    RETURN_IF_NOT_MAIN_THREAD(return origin(item, blockPos, face, clickPos, block, isFirstEvent));
+    RETURN_IF_NOT_MAIN_THREAD(return origin(item, at, face, hit, handSlot, targetBlock, isFirstEvent));
     auto& containerOpenManager  = ContainerOpenManager::getInstance();
     containerOpenManager.player = &this->mPlayer;
-    auto res                    = origin(item, blockPos, face, clickPos, block, isFirstEvent);
+    auto res                    = origin(item, at, face, hit, handSlot, targetBlock, isFirstEvent);
     containerOpenManager.player = nullptr;
     return res;
 }